Fix: getRelativeDate should ignore timezone offset
This function assumes reliable metadata for timezone, but the fact is, neither StashDB or TMDB offer this precision reliably.  StashDB stores this as 00:00:00 UTC, so Whisparr will happily round that backwards to "yesterday" if the user's browser is set to anything less than UTC-00:00.  This change ignores that so that users will see the same dates in both StashDB and Whisparr.  If the datetime DOES include an offset, we'll use it.

(cherry picked from commit fdbe1698da57974782415e727778c0f9d8932064)