Fix: Parser fails when performer is missing alias
When a performer is set on a scene, and the release uses their alias, but the StashDB entry doesn't have it, parser fails to match.  As a conservative fallback, we will see if the parsed title contains the scene cleanTitle, and consider it a match.  At that point in the parsig, we've already narrowed it down to studio and releae date, so this only applies after that.